+++ src/saihubot-cli-skill-eth.js
@@ -3,7 +3,7 @@
 import {t} from 'saihubot-cli-adapter/dist/i18n';
 
 import {getConfig, getRandomItem} from './utils';
-import {i18nValidator, i18nAddr} from './i18n';
+import {i18nAddr} from './i18n';
 import {rpcLastBlock} from './ethRpc';
 
 const isAddr = data => data.length === 42;
@@ -390,42 +390,6 @@ export const skillSearchAnyblock = {
   },
 };
 
-// Side Chain
-
-/**
- * Check address or tx on bscscan.
- *
- * can pass the address, or pre-define the
- * SAIHUBOT_ADDR environment variable
- */
-export const skillSearchBSCscan = {
-  name: 'bscscan',
-  help: '🏦bsc|bscscan [address|tx] - check address or tx on Binance Smart Chain',
-  requirements: {
-    addons: ['search'],
-  },
-  rule: /(^bscscan |^bsc )(.*)|^bscscan$|^bsc$/i,
-  action: function(robot, msg) {
-    let addr = '';
-    if (msg[2] === undefined) {
-      addr = getConfig('ADDR', '');
-      if (addr === '') {
-        robot.send(t('needAddr', {i18n: i18nAddr}));
-        robot.render();
-        return;
-      }
-    }
-    const data = addr || msg[2];
-    if(isAddr(data)) {
-      const url = 'https://bscscan.com/address/' + data;
-      robot.addons.search('Check', data, url, 'Binance Smart Chain');
-    } else {
-      const url = 'https://bscscan.com/tx/' + data;
-      robot.addons.search('Check tx', data, url, 'Binance Smart Chain');
-    }
-  },
-};
-
 // ==== Account/DeFi Balance Tracking ===
 
 /**
@@ -588,14 +552,10 @@ export const skillsAccount = [
   skillSearchZapper,
   skillSearchZerion,
 ];
-export const skillsSideChain = [
-  skillSearchBSCscan,
-];
 
 const skills = [
   skillLastBlock,
   ...skillsAccount,
   ...skillsAddress,
-  ...skillsSideChain,
 ];
 export {skills};